refactor: 경로 및 이름 변경
