+++ src/components/LowerBody/PickedCard/index.jsx
